[level 2] Title: n^2 배열 자르기, Time: 4.57 ms, Memory: 113 MB -BaekjoonHub

+++ 프로그래머스/2/87390. n＾2 배열 자르기/n＾2 배열 자르기.java
@@ -2,7 +2,7 @@ import java.util.*;
 
 class Solution {
     public int[] solution(int n, long left, long right) {
-        int[] answer = new int[(int)right - (int)left + 1];
+        int[] answer = new int[(int)(right - left) + 1];
         
         int index = 0;
         for(long i = left; i <= right; i++) {
